requirements.txt
requests
gradio
Pillow
numpy